modularizes error catching and corrects typo in mult

+++ table.c
@@ -25,9 +25,9 @@ struct table *nova_tabela(){
 
 void add_entrada(struct table *tabela, struct entry *entrada){
     if (tabela != NULL && entrada != NULL){
-        int len = snprintf(NULL,0,"%ld",sizeof(int)*tabela->numero_de_entradas);
+        int len = snprintf(NULL,0,"%d",(int)sizeof(int)*tabela->numero_de_entradas);
         entrada->deslocamento = calloc(1,sizeof(char)*len);
-        sprintf(entrada->deslocamento, "%ld", sizeof(int)*tabela->numero_de_entradas);
+        sprintf(entrada->deslocamento, "%d", (int)sizeof(int)*tabela->numero_de_entradas);
         tabela->numero_de_entradas++;
         tabela->entradas = realloc(tabela->entradas, tabela->numero_de_entradas * sizeof(struct entry));
         tabela->entradas[tabela->numero_de_entradas - 1] = entrada;
@@ -142,3 +142,4 @@ struct entry *search_pilha(struct table_stack *pilha, char *valor){
         }
     } else {return NULL;}
 }
+
